Changes from Chris' Comments
I am certainly using the term minifying loosely, but when I refer to front end files I am including gzip as a form of minifying file size. Also, image compression (gzip or otherwise) is why I kept the term super open to interpretation. The goal being that the student will always be trying to compress or shrink any file that is coming to the browser. I can say for certain that a vast majority of the students in the program will not appreciate the nuances.

Visibility:hidden - Yes. Good suggestion.